refactor(channels): one folder per channel (telegram/, github/)
Move each channel's runtime into its own folder so a channel is a self-contained unit —
the cohesion boundary and the future package boundary (telegram is slated to become a
sibling @kid7st/fastagent-telegram; extracting it should be moving one folder). This also
makes room for each channel's scaffold kit to live beside its runtime in a follow-up.

- src/channels/telegram.ts + telegram-api.ts → src/channels/telegram/
- src/channels/github.ts → src/channels/github/
- http.ts / body.ts / respond.ts stay flat: the invoke handler and shared helpers are not
  channel units.

Pure move. The public export subpaths (./telegram, ./github) are unaffected — they resolve
through the src/telegram.ts / src/github.ts facades, whose one import line now points into
the folder. Internal relative imports shift a level; tests import via the facades, so they
are untouched. typecheck/lint/248 tests green; both subpaths still export as before.
