+++ core/src/github.ts
@@ -4,7 +4,7 @@ export {
   type GithubChannelOptions,
   type GithubEvent,
   type Intent,
-} from "./channels/github.ts";
+} from "./channels/github/github.ts";
 // `GithubEvent.payload` is typed as Schema (the official octokit union); re-exported so `on()`
 // authors can name it without importing @octokit/webhooks-types.
 export type { Schema } from "@octokit/webhooks-types";
